refactor: separate cleanup rule authoring concerns
Keep dialog state, hydration, validation, and payload construction in the authoring shell while extracting the static rule catalog and presentational parameter controls into acyclic modules. Preserve incognito handling and all rule variants. All 628 web tests, web typecheck, and ESLint pass with only existing warnings.
